+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased]
+## Fixed
+- `setFocus` aborts when no valid focus key can be located
+- `smartNavigate` avoids passing `undefined` to `setFocus` when there is no force-focusable key to be found
+
 # [2.3.0]
 ## Added
 - New callback `onArrowRelease`